+++ xerparser/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.5.6"
+__version__ = "0.5.7"
 
 from xerparser.src.parser import xer_to_dict
 from xerparser.src.xer import Xer